metrics: add simple statistics interface
Split out the standard host pool implementation to its own file to make the
hostpool interface more obvious as well.
