+++ error_budget_test.go
@@ -1,7 +1,7 @@
 package hostpool
 
 import (
-	"github.com/bmizerany/assert"
+	"github.com/stretchr/testify/assert"
 
 	"testing"
 	"time"
